build is clean for both projects. Test and Client need to be updated
