+++ Hqub.MusicBrainze.API/Hqub.MusicBrainze.API/Entities/Collections/BaseList.cs
