+++ Hqub.MusicBrainze.API/Hqub.MusicBrainze.API/Entities/Collections/RecordingList.cs
@@ -8,12 +8,9 @@ using System.Xml.Serialization;
 namespace Hqub.MusicBrainz.API.Entities.Collections
 {
     [XmlRoot("recording-list", Namespace = "http://musicbrainz.org/ns/mmd-2.0#")]
-    public class RecordingList : List<Recording>
+    public class RecordingList : BaseList
     {
-        [XmlAttribute("count")]
-        public int QueryCount { get; set; }
-
-        [XmlAttribute("offset")]
-        public int QueryOffset { get; set; }
+        [XmlElement("recording")]
+        public List<Recording> Items { get; set; }
     }
 }
