--- Hqub.MusicBrainze.API/MusicBrainzWebService/Entities/Collection.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Hqub.MusicBrainz.API.Entities
-{
-    public class Collection : Entity
-    {
-
-    }
-}
