--- Hqub.MusicBrainze.API/MusicBrainzWebService/Entities/Metadata.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Hqub.MusicBrainz.API.Entities
-{
-	public abstract class MetadataWrapper : Entity
-	{
-
-	}
-}
