--- Hqub.MusicBrainze.API/MusicBrainzWebService/Entities/Metadata/RecordingMetadataWrapper.cs
@@ -1,41 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Xml.Linq;
-using System.Xml.Serialization;
-using Hqub.MusicBrainz.API.Entities.Collections;
-
-namespace Hqub.MusicBrainz.API.Entities.Metadata
-{
-    [XmlType(Namespace = "http://musicbrainz.org/ns/mmd-2.0#")]
-    [XmlRoot("metadata", Namespace = "http://musicbrainz.org/ns/mmd-2.0#")]
-    public class RecordingMetadataWrapper : MetadataWrapper
-    {
-        public RecordingMetadataWrapper()
-        {
-            Collection = new RecordingList();
-        }
-
-        [XmlElement("recording-list")]
-        public RecordingList Collection { get; set; }
-
-        public override void SetSchema(XElement schema)
-        {
-            base.SetSchema(schema);
-
-            var firstNode = schema.Elements().FirstOrDefault();
-
-            if (firstNode == null)
-                return;
-
-            var countAttribute = firstNode.Attribute("count");
-            if(countAttribute != null)
-            {
-                int count;
-                Collection.QueryCount = int.TryParse(countAttribute.Value, out count) ? count : 0;
-            }
-        }
-    }
-}
