--- Hqub.MusicBrainze.API/MusicBrainzWebService/Entities/Metadata/ReleaseMetadataWrapper.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Xml.Serialization;
-
-namespace Hqub.MusicBrainz.API.Entities.Metadata
-{
-    [XmlType(Namespace = "http://musicbrainz.org/ns/mmd-2.0#")]
-    [XmlRoot("metadata", Namespace = "http://musicbrainz.org/ns/mmd-2.0#")]
-    public class ReleaseMetadataWrapper : MetadataWrapper
-    {
-        [XmlElement("release-list")]
-        public Collections.ReleaseList Collection { get; set; }
-    }
-}
