--- Hqub.MusicBrainze.API/MusicBrainzWebService/Entities/ReleaseGroup.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Xml.Serialization;
-using Hqub.MusicBrainz.API.Entities.Collections;
-
-namespace Hqub.MusicBrainz.API.Entities
-{
-    [XmlType(Namespace = "http://musicbrainz.org/ns/mmd-2.0#")]
-    [XmlRoot("release-group", Namespace = "http://musicbrainz.org/ns/mmd-2.0#")]
-    public class ReleaseGroup : Entity
-    {
-        [XmlAttribute("type")]
-        public string ReleaseGroupType { get; set; }
-
-        [XmlAttribute("id")]
-        public string Id { get; set; }
-
-        [XmlElement("title")]
-        public string Title { get; set; }
-
-        [XmlElement("first-release-date")]
-        public string FirstReleaseDate { get; set; }
-
-        [XmlElement("primary-type")]
-        public string PrimaryType { get; set; }
-
-        [XmlElement("rating")]
-        public Rating Rating { get; set; }
-
-        [XmlElement("tag-list")]
-        public TagList Tags { get; set; }
-    }
-}
