+++ Hqub.MusicBrainze.API/MusicBrainzWebService/MusicBrainzWebService.csproj
@@ -34,18 +34,69 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <!-- A reference to the entire .NET Framework is automatically included -->
-  </ItemGroup>
-  <ItemGroup>
     <TargetPlatform Include="WindowsPhoneApp, Version=8.1" />
     <TargetPlatform Include="Windows, Version=8.1" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Artist.cs">
+      <Link>Entities\Artist.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Collections\ArtistList.cs">
+      <Link>Entities\Collections\ArtistList.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Collections\BaseList.cs">
+      <Link>Entities\Collections\BaseList.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Collections\MediumList.cs">
+      <Link>Entities\Collections\MediumList.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Collections\RecordingList.cs">
+      <Link>Entities\Collections\RecordingList.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Collections\ReleaseGroupList.cs">
+      <Link>Entities\Collections\ReleaseGroupList.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Collections\ReleaseList.cs">
+      <Link>Entities\Collections\ReleaseList.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Collections\TagList.cs">
+      <Link>Entities\Collections\TagList.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Collections\TrackList.cs">
+      <Link>Entities\Collections\TrackList.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Collections\WorkList.cs">
+      <Link>Entities\Collections\WorkList.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\CoverArtArchive.cs">
+      <Link>Entities\CoverArtArchive.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Entity.cs">
+      <Link>Entities\Entity.cs</Link>
+    </Compile>
     <Compile Include="..\Hqub.MusicBrainze.API\Entities\Label.cs">
       <Link>Entities\Label.cs</Link>
     </Compile>
     <Compile Include="..\Hqub.MusicBrainze.API\Entities\LabelInfo.cs">
       <Link>Entities\LabelInfo.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Medium.cs">
+      <Link>Entities\Medium.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Metadata.cs">
+      <Link>Entities\Metadata.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Metadata\ArtistMetadataWrapper.cs">
+      <Link>Entities\Metadata\ArtistMetadataWrapper.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Metadata\RecordingMetadataWrapper.cs">
+      <Link>Entities\Metadata\RecordingMetadataWrapper.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Metadata\ReleaseGroupMetadataWrapper.cs">
+      <Link>Entities\Metadata\ReleaseGroupMetadataWrapper.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Metadata\ReleaseMetadataWrapper.cs">
+      <Link>Entities\Metadata\ReleaseMetadataWrapper.cs</Link>
     </Compile>
     <Compile Include="..\Hqub.MusicBrainze.API\Entities\NameCredit.cs">
       <Link>Entities\NameCredit.cs</Link>
@@ -53,34 +104,29 @@
     <Compile Include="..\Hqub.MusicBrainze.API\Entities\Rating.cs">
       <Link>Entities\Rating.cs</Link>
     </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Recording.cs">
+      <Link>Entities\Recording.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Release.cs">
+      <Link>Entities\Release.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\ReleaseGroup.cs">
+      <Link>Entities\ReleaseGroup.cs</Link>
+    </Compile>
     <Compile Include="..\Hqub.MusicBrainze.API\Entities\Tag.cs">
       <Link>Entities\Tag.cs</Link>
+    </Compile>
+    <Compile Include="..\Hqub.MusicBrainze.API\Entities\Track.cs">
+      <Link>Entities\Track.cs</Link>
     </Compile>
     <Compile Include="..\Hqub.MusicBrainze.API\Entities\Work.cs">
       <Link>Entities\Work.cs</Link>
     </Compile>
-    <Compile Include="Entities\Artist.cs" />
-    <Compile Include="Entities\Collection.cs" />
-    <Compile Include="Entities\Collections\ArtistList.cs" />
-    <Compile Include="Entities\Collections\BaseList.cs" />
-    <Compile Include="Entities\Collections\MediumList.cs" />
-    <Compile Include="Entities\Collections\RecordingList.cs" />
-    <Compile Include="Entities\Collections\ReleaseGroupList.cs" />
-    <Compile Include="Entities\Collections\ReleaseList.cs" />
-    <Compile Include="Entities\Collections\TagList.cs" />
-    <Compile Include="Entities\Collections\TrackList.cs" />
-    <Compile Include="Entities\Collections\WorkList.cs" />
-    <Compile Include="Entities\CoverArtArchive.cs" />
-    <Compile Include="Entities\Medium.cs" />
-    <Compile Include="Entities\Metadata.cs" />
-    <Compile Include="Entities\Metadata\ArtistMetadataWrapper.cs" />
-    <Compile Include="Entities\Metadata\RecordingMetadataWrapper.cs" />
-    <Compile Include="Entities\Metadata\ReleaseMetadataWrapper.cs" />
-    <Compile Include="Entities\Recording.cs" />
-    <Compile Include="Entities\Release.cs" />
-    <Compile Include="Entities\ReleaseGroup.cs" />
-    <Compile Include="Entities\Track.cs" />
-    <Compile Include="Entities\Entity.cs" />
+    <Compile Include="..\Hqub.MusicBrainze.API\QueryParameters.cs">
+      <Link>QueryParameters.cs</Link>
+    </Compile>
+    <Compile Include="Resources\Constants.cs" />
+    <Compile Include="Resources\Messages.cs" />
     <Compile Include="HttpClientException.cs" />
     <Compile Include="ImageLoader.cs" />
     <Compile Include="MyHttpClient.cs" />
